refactor: migrate attack resolution logic to AttackResolver (#1204)
- AttackResolver now holds full PerformPlayerAttack, RollDodge,
  RollPlayerDodge, and RollCrit implementations (was stubs)
- Added PassiveEffectProcessor, DifficultySettings, and shared
  List<CombatTurn> to AttackResolver constructor
- Added CombatTurn property (set per-turn by CombatEngine) and
  SetStats() method (called at combat start) to IAttackResolver
- CombatEngine delegates all four methods to _attackResolver
- Zero behavior change — all 1757 tests pass
